refactor(menu): split executor_admin_users.go by admin sub-function (pure code movement)
- executor_admin_users.go keeps user list/browse UI: runListUsers, lightbar
  browser, admin formatting helpers, USEREDITOR/pending-notice/toggle entry points
- executor_admin_user_editor.go: applyPendingUserChanges, userEditorConfig,
  runUserEditor (shared interactive editor; single ~750-line function keeps
  this file above the 600-line guideline)
- executor_admin_validation.go: validate/unvalidate/ban/delete/purge actions

No behavior change.
